cf: Move DEST assignment
The DEST variable depends on the sysname, so move the assignment of the
DEST variable to the end of the OPENAFS_SYSNAME macro which is where the
sysname value is determined.

Also fix the test for an empty DEST variable, which put the dummy 'x'
after the variable, not before, to guard against values that start with
a dash.

Change-Id: Iaf8deefc685a6c9b9897d83b25b8919792847796
Reviewed-on: https://gerrit.openafs.org/15821
Reviewed-by: Mark Vitale <mvitale@sinenomine.net>
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Reviewed-by: Cheyenne Wills <cwills@sinenomine.net>
Reviewed-by: Michael Meffie <mmeffie@sinenomine.net>