bozo: Make nextRestart, nextDay local vars
The global variables nextRestart and nextDay are only referenced in
BozoDaemon(), a function that never returns. There is no reason for
these to be globals, so move them to be local variables to make it
obvious that there are no concerns with locking, and no possible issues
with accessing these vars from different functions or different threads.

Change-Id: I38a8fc2227455dc3d7bc2899aa20cfa5475b4ad4
Reviewed-on: https://gerrit.openafs.org/15835
Reviewed-by: Benjamin Kaduk <kaduk@mit.edu>
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Reviewed-by: Michael Meffie <mmeffie@sinenomine.net>