serverLog: call setvbuf unconditionally
The setvbuf() function is part of C89, and we require C89 support
from our platforms.  Remove the configure check and the corresponding
conditional logic.  (As it happens, we were calling setvbuf() without
the conditional elsewhere anyway, via the setlinebuf() macros, so
we should not see any additional breakage from this change.)

Change-Id: Iff850bfbafdb9a17f50410d309d9d9e9be3fa951
Reviewed-on: https://gerrit.openafs.org/15841
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Reviewed-by: Michael Meffie <mmeffie@sinenomine.net>
Reviewed-by: Cheyenne Wills <cwills@sinenomine.net>