vos: Avoid 'vos zap' warning on unrelated VL sites
vos zap currently prints a warning message when a VLDB entry is found
for the volume being zapped:

    $ vos zap fs.example.com vicepa 1234
    Warning: Entry for volume number 1234 exists in VLDB (but we're zapping it anyway!)

However, vos only checks whether the VLDB entry exists at all, and
doesn't check the location being zapped. If the VLDB entry exists, but
it does not reference the specific server and partition being zapped,
the warning message is still printed, which is confusing.

Update vos zap to print the warning message only when zapping a location
that exists in the VLDB entry, and to print the location we checked. The
warning now looks like:

    $ vos zap fs.example.com vicepa 1234
    Warning: Entry for volume 1234 on server fs.example.com /vicepa exists in VLDB (but we're zapping it anyway!)

Change-Id: I382d9b3a5960deb06e7774b8ab6a8b36b92cc0a6
Reviewed-on: https://gerrit.openafs.org/15849
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Reviewed-by: Cheyenne Wills <cwills@sinenomine.net>
Reviewed-by: Andrew Deason <adeason@sinenomine.net>