rxkad: rename ENCRYPT macro to FCRYPT_ENCRYPT
rxkad/fcrypt defines 2 macros: ENCRYPT and DECRYPT.  The names are
pretty generic and do not clearly associate with fcrypt.

In addition, the name ENCRYPT collide with the Microsoft provided header
'ntddndis.h' that uses ENCRYPT as an ENUM (on newer versions of the
Windows development kits):

    typedef enum    _OFFLOAD_OPERATION_E
    {
        AUTHENTICATE = 1,
        ENCRYPT
    }

This results in a build error on Windows when fcrypt.h is included with
ntddndis.h:
    ... ntddndis.h(2212): error C2059: syntax error: 'constant'

Rename the ENCRYPT/DECRYPT macros to FCRYPT_ENCRYPT/FCRYPT_DECRYPT in
fcrypt.h to a name that relates the macros back to fcrypt.

Note: The ENCRYPT/DECRYPT symbols are part of a public interface
installed in fcrypt.h, but keeping the old names are impractical, so we
are changing them anyway.

Change-Id: I9c51c81fab7fcbf0dae01569852ca94c0e6a0439
Reviewed-on: https://gerrit.openafs.org/15868
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Reviewed-by: Michael Meffie <mmeffie@sinenomine.net>
Reviewed-by: Benjamin Kaduk <kaduk@mit.edu>