Linux CM: Fix leak of group_info on setpag()
Commit ee2fbffb04bb ("LINUX: Revert group changes on keyring
failure") made __setpag return a pointer (with a ref) to the old
set of groups so that it could be reinstated in case of error.
But it only put back the reference on error, which leads to a
leak of 16 bytes for every successful setpag() call.

Always put back the ref to the old group_info, whether it is a
failure or a success.

Change-Id: I9805838a7bda3b73c38c69b3378c35b8ba1b5fa9
Reviewed-on: https://gerrit.openafs.org/15899
Reviewed-by: Cheyenne Wills <cwills@sinenomine.net>
Reviewed-by: Michael Meffie <mmeffie@sinenomine.net>
Tested-by: BuildBot <buildbot@rampaginggeek.com>