macos: Set 'ret' in AFSBackgrounder's renewTicket
On macOS 14 (Sonoma):

If the AFSBackgrounder is running and enabled for automatic renewal of
Kerberos tickets, AND if there is no current Kerberos ticket, then the
renewal task will 1) fail to prompt for a Kerberos principal and
password and 2) throw an exception and crash.

Because the launchd task will automatically restart AFSBackgrounder, it
will crash repeatedly until either a ticket is manually obtained (via
kinit) or the AFSBackgrounder is disabled (by unchecking the
Backgrounder checkbox in the OpenAFS preference pane), or the prefpane
'Enable auto-renew' box is unchecked.

The root cause for this crash is undefined behavior in renewTicket():
since commit 62cccc4547bd (macos: avoid KLRenewInitialTickets crash in
Lion), stack variable 'ret' is never assigned a value before we test it
for 0.

In previous versions of macOS (10.7 (Lion) through 13 (Ventura)), this
code has worked as expected, prompting the user for a Kerberos principal
and password.  However, with the version of Xcode (Apple clang version
15.0.0 (clang-1500.1.0.2.5)) introduced by macOS 14 (Sonoma), the
compiler elides all code after the errant 'if (ret == 0)', including
the Kerberos prompt logic (krb5_cc_get_principal).

Mitigate the undefined behavior by ensuring 'ret' has a defined value
before it is tested. Set it to the error code of krb5_cc_default(), so
we won't do anything if krb5_cc_default() fails.

Reviewed-on: https://gerrit.openafs.org/15971
Reviewed-by: Mark Vitale <mvitale@sinenomine.net>
Reviewed-by: Michael Meffie <mmeffie@sinenomine.net>
Reviewed-by: Andrew Deason <adeason@sinenomine.net>
Tested-by: Andrew Deason <adeason@sinenomine.net>
(cherry picked from commit 9585f92d785779ae1d0e798f81b896743c1236bd)

Change-Id: I9b1844ad87d427e8c6b29da36bc31b2d2fc13e50
Reviewed-on: https://gerrit.openafs.org/16260
Reviewed-by: Andrew Deason <adeason@sinenomine.net>
Reviewed-by: Michael Meffie <mmeffie@sinenomine.net>
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Reviewed-by: Benjamin Kaduk <kaduk@mit.edu>