sys: Add lpioctl to liboafs_sys.la.sym
Since OpenAFS 1.0, src/venus/test/getinitparams has contained a
reference to 'lpioctl', but this is not exported by liboafs_sys.la.sym.
This causes the build to fail with --disable-static:

    $ make V=0
      LD  .../src/venus/test/getinitparams
    /usr/bin/ld: getinitparams.o: in function `GetInitParamsCmd':
    .../src/venus/test/getinitparams.c:75: undefined reference to `lpioctl'
    collect2: error: ld returned 1 exit status

    FAILED COMMAND: /bin/bash ../../../libtool --quiet --mode=link --tag=CC gcc [...] -o getinitparams getinitparams.o .../src/sys/liboafs_sys.la [...]

Add the missing symbol to liboafs_sys.la.sym. Note that lpioctl() isn't
a function on AIX (it's a syscall), but on all other platforms it is a
function provided by libsys.

Change-Id: Ie54c46a8eb1de92698472636860186fe274daf1c
Reviewed-on: https://gerrit.openafs.org/15994
Reviewed-by: Cheyenne Wills <cwills@sinenomine.net>
Reviewed-by: Michael Meffie <mmeffie@sinenomine.net>
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Reviewed-by: Ben Huntsman <ben@huntsmans.net>
Tested-by: Ben Huntsman <ben@huntsmans.net>